+++ src/resources/manifest.php
@@ -4,5 +4,5 @@ declare(strict_types=1);
 
 return [
     'name' => 'AppShell',
-    'version' => '4.4.0'
+    'version' => '4.5.0'
 ];