+++ Changelog.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 ##### 2024-XX-YY
+
 - Added Customer purchases chart to the Customer show page
+- Changed the minimal Customer library dependency to v3.1
 
 ## 4.4.0
 ##### 2024-11-23